+++ src/main/kotlin/blackjack/model/deck/Card.kt
@@ -1,6 +1,3 @@
 package blackjack.model.deck
 
-import blackjack.util.CardNumber
-import blackjack.util.Shape
-
 data class Card(val cardNumber: CardNumber, val shape: Shape)
