+++ src/main/kotlin/blackjack/model/deck/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack.util
+package blackjack.model.deck
 
 enum class CardNumber(val value: String, val score: Int) {
     TWO("2", 2),
