+++ src/main/kotlin/blackjack/model/participant/state/Blackjack.kt
@@ -1,6 +1,6 @@
 package blackjack.model.participant.state
 
-import blackjack.util.CompetitionResult
+import blackjack.model.participant.CompetitionResult
 
 class Blackjack : ParticipantState {
     override fun getResult(opponentScore: Int): CompetitionResult {
