+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,12 +1,12 @@
 package blackjack
 
 import blackjack.model.deck.Deck
+import blackjack.model.participant.CompetitionResult
 import blackjack.model.participant.Dealer
 import blackjack.model.participant.Players
 import blackjack.testmachine.BlackjackCardMachine
 import blackjack.testmachine.BustCardMachine
 import blackjack.testmachine.NormalCardMachine
-import blackjack.util.CompetitionResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
