+++ src/test/kotlin/blackjack/testmachine/BustCardMachine.kt
@@ -2,8 +2,8 @@ package blackjack.testmachine
 
 import blackjack.model.deck.Card
 import blackjack.model.deck.CardMachine
-import blackjack.util.CardNumber
-import blackjack.util.Shape
+import blackjack.model.deck.CardNumber
+import blackjack.model.deck.Shape
 
 class BustCardMachine : CardMachine {
     override fun handle(cards: List<Card>): List<Card> {
